cloud_run/app/requirements.txt
Flask~=3.1
gunicorn~=23.0
google-genai>=1.0
